Removed old Saas folder & revisit code in delete by url

+++ inc/Engine/Optimization/RUCSS/Controller/UsedCSS.php
@@ -102,6 +102,13 @@ class UsedCSS {
 		return false;
 	}
 
+	/**
+	 * Delete used css based on URL.
+	 *
+	 * @param string $url The page URL.
+	 *
+	 * @return boolean
+	 */
 	public function delete_used_css( string $url ) : bool {
 		$used_css_arr = $this->used_css_query->query( [ 'url' => $url ] );
 
@@ -112,6 +119,9 @@ class UsedCSS {
 		$deleted = true;
 
 		foreach ( $used_css_arr as $used_css ) {
+			if ( empty( $used_css->id ) ) {
+				continue;
+			}
 			$deleted = $deleted && $this->used_css_query->delete_item( $used_css->id );
 		}
